Inline NllTypeRelating into its only usage site
